Update Sonata GPIO driver to reflect HW changes
Sonata's hardware has changed with respect to the GPIO since the GPIO
driver was last updated - this was also out of date for the 0.4 and
0.4.1 releases, as the GPIO HW change had already occured at that point.

This means that attempts to access debounced input or use any of the
other GPIO instances (raspberry PI HAT, Arduino Shield and PMOD) are not
currently working.

This commit updates Sonata's GPIO driver, whilst still retaining
backwards compatibility for all interfaces that were working at the time
of the 0.4.1 release (writing to LEDs, reading from DIP switches, and
using the general GPIO's `input` and `output` registers). Now a
separate struct is defined for each GPIO instance, with different masks
that correspond to the actual GPIO implementation of that instance.

Users of the driver can now also access the `output_enable` and
`debouncedInput` registers, allowing non-general GPIO to be used as
either an output or an input, and access to debounced input.