Fix incorrect definitions & docs in USBDev Interrupts
This commit fixes a bug in the USBDev interrupts where the Available Out
Buffer empty and Receive Fifo Full were defined the wrong way around by
accident.

It also updates an incorrect comment saying that a FIFO is full for an
interrupt, when the interrupt is actually for the FIFO being non-empty.