Implement Pinmux driver for Sonata
Implements a basic Pinmux driver for Sonata, with a small change to the
latest autogenerated version for the 1.0 mappings found in Sonata
System.

This may require more work to match code style standards but this driver
should allow users to interact with Sonata's Pinmux without having to
wait for potential refactoring.